Meta

+++ framework-web/deploy/mysql/db/equity_orders.sql
@@ -2,15 +2,15 @@ DROP TABLE IF EXISTS equity_orders;
 
 CREATE TABLE equity_orders
   ( 
-     order_identifier                        BIGINT, 
+     order_identifier                        INT(10), 
      order_event_number                      INT(2), 
      order_placement_date                    VARCHAR(50), 
-     parent_order_id                         BIGINT, 
+     parent_order_id                         INT(10), 
      parent_order_placement_date             VARCHAR(50), 
      root_order_identifier                   VARCHAR(50), 
      root_order_placement_date               VARCHAR(50), 
      order_allocation_relationship_code      VARCHAR(50), 
-     allocated_order_id                      BIGINT, 
+     allocated_order_id                      INT(10), 
      allocated_order_placement_date          VARCHAR(50), 
      order_event_type_cd                     VARCHAR(50), 
      order_event_date                        VARCHAR(50),
